crc: See https://github.com/msgpack/msgpack/pull/114
_msgpack_pack_crc_user must be user-defined to map from,to from the encoding context
se empty reserved msgpack 0xc6 type for a crc32 as uint32 at the end of the buffer.
See msgpack-perl for an added add_crc() pack method.

Optional CRC tag for basic security and data corruption.
If pack provides a crc checksum, unpack should check the result against the given
crc and report an error otherwise.
The crc tag must be the last tag in the buffer.
The crc checksum does not include itself and its type tag.